+++ Equilibrium/Component/DamageStackMono.cs
@@ -39,7 +39,9 @@ namespace Equilibrium.Component
         {
             var proj = bullet.GetComponent<ProjectileHit>();
             if (proj != null)
+            {
                 proj.damage *= currentMultiplier;
+            }
             currentMultiplier += (increment * incrementIncrement);
         }